[Fix] Firefly color renders white instead of tinted
EntityFirefly.makeColor cast each channel through (byte) before handing
it to ColorUtil.color, which packs ARGB with (r << 16) | (g << 8) | b.
Channel values above 127 became negative signed bytes, and once
sign-extended and OR-ed together they forced the red and green channels
to 0xFF - so every firefly came out white/pale yellow regardless of hue.

This stayed hidden until the 1.21 migration switched makeColor's result
from float channels (red / 255F) to the packed-int ColorUtil.color path,
which is when the latent byte overflow started producing white.

Cast the channels with (int) instead of (byte) so values stay in 0-255
and pack cleanly, restoring each firefly's random glowing hue. The render
pipeline and color plumbing were already correct.

+++ src/main/java/org/betterx/betternether/entity/EntityFirefly.java
@@ -471,34 +471,34 @@ public class EntityFirefly extends DespawnableAnimal implements FlyingAnimal {
         float f7 = brightness * (1.0F - saturation * (1.0F - f4));
         switch ((int) f3) {
             case 0:
-                red = (byte) (brightness * 255F + 0.5F);
-                green = (byte) (f7 * 255F + 0.5F);
-                blue = (byte) (f5 * 255F + 0.5F);
+                red = (int) (brightness * 255F + 0.5F);
+                green = (int) (f7 * 255F + 0.5F);
+                blue = (int) (f5 * 255F + 0.5F);
                 break;
             case 1:
-                red = (byte) (f6 * 255F + 0.5F);
-                green = (byte) (brightness * 255F + 0.5F);
-                blue = (byte) (f5 * 255F + 0.5F);
+                red = (int) (f6 * 255F + 0.5F);
+                green = (int) (brightness * 255F + 0.5F);
+                blue = (int) (f5 * 255F + 0.5F);
                 break;
             case 2:
-                red = (byte) (f5 * 255F + 0.5F);
-                green = (byte) (brightness * 255F + 0.5F);
-                blue = (byte) (f7 * 255F + 0.5F);
+                red = (int) (f5 * 255F + 0.5F);
+                green = (int) (brightness * 255F + 0.5F);
+                blue = (int) (f7 * 255F + 0.5F);
                 break;
             case 3:
-                red = (byte) (f5 * 255F + 0.5F);
-                green = (byte) (f6 * 255F + 0.5F);
-                blue = (byte) (brightness * 255F + 0.5F);
+                red = (int) (f5 * 255F + 0.5F);
+                green = (int) (f6 * 255F + 0.5F);
+                blue = (int) (brightness * 255F + 0.5F);
                 break;
             case 4:
-                red = (byte) (f7 * 255F + 0.5F);
-                green = (byte) (f5 * 255F + 0.5F);
-                blue = (byte) (brightness * 255F + 0.5F);
+                red = (int) (f7 * 255F + 0.5F);
+                green = (int) (f5 * 255F + 0.5F);
+                blue = (int) (brightness * 255F + 0.5F);
                 break;
             case 5:
-                red = (byte) (brightness * 255F + 0.5F);
-                green = (byte) (f5 * 255F + 0.5F);
-                blue = (byte) (f6 * 255F + 0.5F);
+                red = (int) (brightness * 255F + 0.5F);
+                green = (int) (f5 * 255F + 0.5F);
+                blue = (int) (f6 * 255F + 0.5F);
                 break;
         }
